Revamp structure to use __main__.py and settings.py
