+++ luke_bot/__init__.py
@@ -1,0 +1,6 @@
+import logging
+import sys
+
+logger = logging.getLogger()
+logger.setLevel(logging.DEBUG)
+logger.addHandler(logging.StreamHandler(sys.stdout))
